requirements.txt
fastapi
uvicorn
requests
jsonschema
dotenv
pillow
